dashboard -> configure, settings -> dashboard
