+++ src/app/pages/configure/components/base.component.ts
