+++ src/app/pages/configure/components/monitoring/monitoring.component.ts
