+++ src/app/pages/configure/components/project-type/project-type.component.ts
