+++ src/app/pages/configure/components/skeleton/skeleton.component.ts
