+++ src/app/pages/configure/components/steps/steps.component.ts
