+++ src/app/pages/configure/components/summary/summary.component.ts
