+++ src/app/pages/configure/components/tech-stack/tech-stack.component.ts
