+++ src/app/routes.ts
@@ -1,9 +1,9 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from '@pages/home/home.component';
-import { DashboardComponent } from '@pages/dashboard/dashboard.component';
 import { ContactComponent } from '@pages/contact/contact.component';
 import { AuthComponent } from './pages/auth/auth.component';
-import { SettingsComponent } from './pages/settings/settings.component';
+import { ConfigureComponent } from './pages/configure/configure.component';
+import { DashboardComponent } from './pages/dashboard/dashboard.component';
 
 const routeConfig: Routes = [
   {
@@ -17,9 +17,9 @@ const routeConfig: Routes = [
     title: 'Couchtec - Auth'
   },
   {
-    path: 'dashboard',
-    component: DashboardComponent,
-    title: 'Couchtec - Dashboard'
+    path: 'configure',
+    component: ConfigureComponent,
+    title: 'Couchtec - Configure'
   },
   {
     path: 'support',
@@ -27,9 +27,9 @@ const routeConfig: Routes = [
     title: 'Couchtec - Support'
   },
   {
-    path: 'settings',
-    component: SettingsComponent,
-    title: 'Couchtec - Settings'
+    path: 'dashboard',
+    component: DashboardComponent,
+    title: 'Couchtec - Dashboard'
   },
   {
     path: '**',
